feat: add emoji
